Move BlockQueue to own file
